requirements.txt
langchain
langchain-anthropic
langchain-community
langgraph
langsmith
simple-salesforce==1.12.4
xmltodict==0.13.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0 